Extract YAML loading and parsing into config module
