+++ tests/integration/testcases.py
@@ -1,6 +1,7 @@
 from __future__ import unicode_literals
 from __future__ import absolute_import
 from compose.service import Service
+from compose.config import make_service_dict
 from compose.cli.docker_client import docker_client
 from compose.progress_stream import stream_output
 from .. import unittest
@@ -21,14 +22,15 @@ class DockerClientTestCase(unittest.TestCase):
                 self.client.remove_image(i)
 
     def create_service(self, name, **kwargs):
+        kwargs['image'] = "busybox:latest"
+
         if 'command' not in kwargs:
             kwargs['command'] = ["/bin/sleep", "300"]
+
         return Service(
             project='composetest',
-            name=name,
             client=self.client,
-            image="busybox:latest",
-            **kwargs
+            **make_service_dict(name, kwargs)
         )
 
     def check_build(self, *args, **kwargs):
